js: fix NativeDate
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk@12470 5f5364db-9458-4db8-a492-e30667be6df6

+++ js/src/main/java/com/gargoylesoftware/js/nashorn/internal/objects/NativeDate.java
@@ -1444,7 +1444,7 @@ public final class NativeDate extends ScriptObject {
                     staticHandle("UTC", double.class, Object.class, Object[].class));
             UTC.setArity(7);
             now = ScriptFunction.createBuiltin("now",
-                    staticHandle("now", long.class, Object.class));
+                    staticHandle("now", double.class, Object.class));
             final Prototype prototype = new Prototype();
             PrototypeObject.setConstructor(prototype, this);
             setPrototype(prototype);
